- Renamed FriendlistAdapter to FriendsAdapter, streamlined its functionality and moved it to a new package, better reflecting its purpose - Renamed GetUserFriendlistPort to CheckIfFriendsPort, now it returns a boolean value, indicating if two users are friends, instead of returning a user's friendlist - Moved the guard clauses regarding inviting users to events to the Event domain object - Moved the guard clauses regarding confirming attendance to the Event domain object - Split test classes for both Event and User domains into multiple classes for better maintainability - Replaced usage of `assertTrue(somecollection.contains(someobject))` with `assertThat(somecollection, contains(someobject))` from Hamcrest
Massive thanks to @krzysztofcybulski for his feedback!
