--- src/main/java/pl/eukon05/eventboard/event/application/port/in/GetUserFriendlistPort.java
@@ -1,7 +1,0 @@
-package pl.eukon05.eventboard.event.application.port.in;
-
-import java.util.List;
-
-public interface GetUserFriendlistPort {
-    List<String> getUserFriendlist(String userID);
-}
